Files in excel folder are deleted before creation of the new ones

+++ src/main/java/MediaPoolMalBridge/service/Bridge/excelcreator/excelfilesserver6_5/BridgeCreateExcelXSSFFileUniqueThreadService.java
@@ -101,9 +101,11 @@ public class BridgeCreateExcelXSSFFileUniqueThreadService extends AbstractBridge
 
     private void deleteOldFilesFromExcelDirectory() {
         final File[] files = new File(appConfig.getExcelDir()).listFiles();
-        for (File file : files) {
-            if (file.isFile()) {
-                file.delete();
+        if (files != null) {
+            for (File file : files) {
+                if (file.isFile() && file.exists()) {
+                    file.delete();
+                }
             }
         }
     }
